+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/scripting/actions/types/ActionFactory.kt
@@ -30,6 +30,7 @@ class ActionFactory {
                 GetVariableActionType(),
                 HashActionType(),
                 HmacActionType(),
+                PlaySoundActionType(),
                 PromptActionType(),
                 RenameShortcutActionType(),
                 SelectionActionType(),